+++ src/index.ts
@@ -1,7 +1,7 @@
 export { MysqlCoreError, MysqlEntityDefinitionError, MysqlEntityNotFound, MysqlEntityValidationError } from "./mysql.error"
-export { MysqlQuery, MysqlStatusCallback, MysqlConnector } from "./mysql.service"
+export { MysqlQuery, MysqlConnectionCallback, MysqlConnector } from "./mysql.service"
 export { entitiesMatch } from "./mysql.utils"
-export { MysqlMigration, MysqlMigrationHandler } from "./mysql.migration"
+export { MysqlMigration, MysqlMigrationHandler, MysqlMigrationStatusCallback } from "./mysql.migration"
 export { QueryCreateTable, QueryDropTable, QueryRenameTable, QueryAlterTable, 
   MysqlOperator, MysqlJoinType, QueryFind, QueryFindOne, QueryCount, QueryInsert, QueryUpdate, QueryDelete } from "./mysql.query"
 export { MysqlDataType, MysqlDefaultValues, MysqlValidationError, MysqlEntity, MysqlParameterValidation } from "./mysql.schema"