+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v2.0.0] 2025-04-08
+### Add
+* Add `onDisconnected` and `onConnected` callbacks for MsqlConnector and `onStatusUpdate` callback for the `MysqlMigrationHandler` tool.
+### Updated
+* Call `super` on lifecycle events.
+
 ## [v1.0.6] 2025-03-14
 ### Fixed
 * Fix typo for number and boolean values in the Entity Constructor method